restructure samplers module
