+++ rust/src/meas/ptr/mod.rs
@@ -13,7 +13,7 @@ use crate::metrics::{IntDistance, L1Distance};
 use crate::measures::{SmoothedMaxDivergence, SMDCurve};
 use crate::domains::{AllDomain, MapDomain};
 use crate::error::Fallible;
-use crate::samplers::SampleLaplace;
+use crate::traits::samplers::SampleLaplace;
 use crate::traits::{CheckNull, InfCast};
 
 // propose-test-release count grouped by unknown categories,
